Plot selection of target trees

plot_target() can be used with a target_inv dataset or the output of the compete_inv function to inspect the spatial positions of the target trees.

Usage

plot_target(inv, radius = NULL)

Arguments

inv

object of class compete_inv or target_inv.

radius

numeric of length 1, Search radius (in m) around target tree wherein all neighboring trees are classified as competitors. Only needed for target_inv objects in methods that are not radius dependent (i.e., in cases where target_source defined by a character or logical vector or in cases where target_source = "all_trees").

Value

A plot of the spatial arrangement of target trees.

Details

The function creates a plot of the trees in the forest inventory dataset where the target trees and the surrounding search radii are highlighted. If they were created with target_source = "buff_edge" or target_source = "exclude_edge", the estimated plot margin (and, in case of "buff_edge", also the buffer to the margin) are added as a polygon.

This function is meant as a visual inspection tool for checking the validity of the choice of target trees.

See also

Examples

if (FALSE) { # \dontrun{
  # plot neighborhood for existing inventory
  comp <- compete_inv(inv, "buff_edge", radius = 10)
  plot_target(comp)
} # }