Quantify size- and distance-dependent competition using inventory data

'compete_inv()' returns a specific distance-height-dependent or distance- dbh-dependent competition index (or group of indices) for all trees within a forest plot or specified target trees.

Usage

compete_inv(
  inv_source,
  target_source = "buff_edge",
  radius,
  method = c("all_methods", "CI_Hegyi", "CI_Braathe", "CI_RK1", "CI_RK2", "CI_RK3",
    "CI_RK4"),
  x = NULL,
  y = NULL,
  dbh = NULL,
  height = NULL,
  id = NULL,
  dbh_unit = c("cm", "m", "mm"),
  height_unit = c("m", "cm", "mm"),
  verbose = TRUE,
  tol = 1,
  ...
)

Arguments

inv_source

either an object of class target_inv, or any object that can be imported by read_inv() (in this case, x, y, id, dbh, and/or height can be specified as in this function – see the corresponding documentation for details). If provided with a target_inv object, the function ignores target_source and overrides further arguments passed to read_inv() and define_target().

target_source

one of the following:

a path to an any object that can be imported by read_inv() (in this case, column specifications have to be the same as in inv_source - if this is not possible, load outside of compete_inv()).

a vector of class "character" containing the tree IDs identifying the target trees in the same format as in the id column of inv,

a vector of class logical specifying for each row of inv whether the corresponding tree is a target tree,

another object of class forest_inv containing the coordinates of the target trees. In this case, the coordinates are matched against the coordinates in inv and IDs may differ (useful e.g. when target trees are defined based on GPS coordinates and matched against an airborne laser scanning dataset).

a character vector of length 1 defining the method by which the target trees should be determined. Allowed are "buff_edge" for excluding all trees that are at least one search radius from the forest edge, "exclude_edge" for only excluding edge trees or "all_trees" for including all trees of the dataset (which is hardly ever a good idea unless all trees in the entire forest are in the dataset). The standard is "buff_edge". See define_target() for details.

radius

numeric of length 1. Search radius (in m) around the target tree. All neighboring trees within this radius are classified as competitors.

method

character string assigning the method for quantifying competition. dbh-distance-dependent methods are "CI_Hegyi", "CI_RK1", and "CI_RK2". Height-distance-dependent methods are "CI_Braathe", "CI_RK3", and "CI_RK4". "all_methods" can be specified to compute all the indices that can be calculated with the available data.

x

character of length 1 or name of the variable in inv_source containing the x coordinates of the tree in m. If NULL (default), the function tries to identify the x coordinate from the data.

y

character of length 1 or name of the variable in inv_source containing the y coordinates of the tree in meters. If NULL (default), the function tries to identify the y coordinate from the data.

dbh

character of length 1 or name of the variable in inv_source containing the diameter at breast height of the tree (by default in cm, but can be defined via dbh_unit). If NULL (default), the function tries to identify the dbh from the data.

height

character of length 1 or name of the variable in inv_source containing the height of the tree (by default in m, but can be defined via heigh_unit). If NULL (default), the function tries to identify the height from the data.

id

character of length 1 or name of the variable in inv_source containing a unique tree ID. If NULL (default), the function tries to identify the ID from the data. If this is not possible, the trees are assigned a unique number. All IDs are coerced to character.

dbh_unit

character of length 1. Unit for the diameter measurements (one of "cm", "m" or "mm". defaults to "cm").

height_unit

character of length 1. Unit for the diameter measurements (one of "m", "cm" or "mm". defaults to "m").

verbose

logical of length 1. Should information about progress be printed? Defaults to TRUE.

tol

numeric of length 1. Tolerance for the match with the tree coordinates. If coordinates are measured in the field with GPS, but inv_source contains x and y coordinates from a larger number of trees obtained by segmentation, this is the tolerance for the matching the forest inventory against the target tree positions. If no forest tree is within the tolerance to a target tree, no competition indices will be calculated for this tree and the function will return a warning. tol defaults to 1 m, but should be chosen depending on the measurement accuracy of the GPS coordinates.

...

additional arguments passed on to data.table::fread().

Value

object of classcompete_inv: a modified data.frame with the position and size of the designated target tree(s) and one or more competition indices depending on chosen method(s).

Details

Using an inventory table to easily quantify distance-dependent tree competition for a list of trees within a plot or all the trees. The input data can either be taken directly from field measurements or derived beforehand from LiDAR point clouds. The function calculates 6 different Competition indices, based on tree heights or dbh and distance to competitors.

Methods

The available competition indices are computed according to the following equations, where \(d_i\) and \(h_i\) are the dbh and height of neighbor tree \(i\), \(d\) and \(h\) are dbh and height of the focal tree, and \(dist_i\) is the distance of neighbor tree \(i\).

Diameter-based competition indices

CI_Hegyi introduced by Hegyi (1974):
\(CI_{Hegyi} = \sum_{i=1}^{n} d_{i} / (d \cdot dist_{i})\)

CI_RK1 according to CI1 Rouvinen & Kuuluvainen (1997):
\(CI_{RK1} = \sum_{i=1}^{n} \mathrm{arctan}(d_{i} / dist_{i})\)

CI_RK2 according to CI3 in Rouvinen & Kuuluvainen (1997):
\(CI_{RK2} =\sum_{i=1}^{n} (d_{i} / d) \cdot \mathrm{arctan}(d_{i} / dist_{i})\)

Height-based competition indices

CI_Braathe according to Braathe (1980):
\(CI_{Braathe} = \sum_{i=1}^{n} h_{i} / (h \cdot dist_{i})\)

CI_RK3 according to CI5 in Rouvinen & Kuuluvainen (1997):
\(CI_{RK3} = \sum_{i=1}^{n} \mathrm{arctan}(h_{i} / dist_{i})\) for all trees with \(h_{i} > h\)

CI_RK4 based on CI3 in Rouvinen & Kuuluvainen (1997) and Contreras et al. (2011):
\(CI_{RK4} = \sum_{i=1}^{n} (h_{i} / h) \cdot \mathrm{arctan}(h_{i} / dist_{i})\)

Tree Segmentation

Various approaches can be used to segment (airborne) laser scanning point clouds into single trees and to obtain inventory data based it. Existing R packages for this are for example:

TreeLS package for automated segmentation of terrestrial/mobile laser scans

lidR package with different options to segment the point cloud or a Canopy Height Model (CHM)

itcLiDARallo within the package itcSegment

Be careful with low resolution/low density point clouds, as oversegmentation of trees is usually an issue!

Literature

Hegyi, F., 1974. A simulation model for managing jackpine stands. In: Fries, J. (Ed.), Proceedings of IUFRO meeting S4.01.04 on Growth models for tree and stand simulation, Royal College of Forestry, Stockholm.

Braathe, P., 1980. Height increment of young single trees in relation to height and distance of neighboring trees. Mitt. Forstl. VersAnst. 130, 43–48.

Rouvinen, S., Kuuluvainen, T., 1997. Structure and asymmetry of tree crowns in relation to local competition in a natural mature Scot pine forest. Can. J. For. Res. 27, 890–902.

Contreras, M.A., Affleck, D. & Chung, W., 2011. Evaluating tree competition indices as predictors of basal area increment in western Montana forests. Forest Ecology and Management, 262(11): 1939-1949.

See also

Examples

if (FALSE) { # \dontrun{
# Quantify the Hegyi index for specified target trees with search radius 10m
CI <- compete_inv("path/to/invtable.csv",
  "path/to/target_trees.csv", radius = 10, method = "CI_Hegyi")
# Quantify the Braathe index for specified target trees with search radius
#10m and adjust
CI <- compete_inv("path/to/invtable.csv",
  "path/to/target_trees.csv", radius = 10, method = "CI_Braathe")
# Specify the units of dbh or height of your input data
CI <- compete_inv("path/to/invtable.csv",
  "path/to/target_trees.csv", radius = 10, method = "CI_Hegyi",
  dbh_unit = "m", height_unit = "m")
# Quantify all available indices for all trees within the plot that are one
#search radius away from plot edge
CI <- compete_inv("path/to/invtable.csv", target_source = "buff_edge",
        radius = 12, method = "all_methods")
} # }